+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [`0.5.0`] Unreleased
 
 ### `nexus-cli`
 
@@ -16,7 +16,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `nexus scheduler occurrence add`
   - `nexus scheduler periodic set` / `disable`
 
-## [`0.4.0`] - 2025-01-07
+### `nexus-sdk`
+
+#### Changed
+
+- leader and crypto caps in PTB templates are now party objects
+- added `ToolRegistryCreated` as tracked event
+
+## [`0.4.0`] - 2026-01-07
 
 ### `nexus-cli`